flow between home screen and survey established
