+++ client/src/survey/components/AnswerChoices.tsx
